Consolidate I/O tests
