+++ tests/test_snps.py
@@ -205,25 +205,6 @@ class TestSnps(BaseSNPsTestCase):
         assert list(s.snps.columns.values) == ["chrom", "pos", "genotype"]
         assert s.snps.index.name == "rsid"
 
-    def generic_snps(self):
-        return self.create_snp_df(
-            rsid=["rs" + str(i) for i in range(1, 9)],
-            chrom=["1"] * 8,
-            pos=list(range(101, 109)),
-            genotype=["AA", "CC", "GG", "TT", np.nan, "GC", "TC", "AT"],
-        )
-
-    def generic_snps_vcf(self):
-        df = self.generic_snps()
-        return df.append(
-            self.create_snp_df(
-                rsid=["rs" + str(i) for i in range(12, 18)],
-                chrom=["1"] * 6,
-                pos=list(range(112, 118)),
-                genotype=[np.nan] * 6,
-            )
-        )
-
     def snps_GRCh38_func(self):
         return self.create_snp_df(
             rsid=["rs3094315", "rsIndelTest", "rs2500347", "rs11928389"],
@@ -239,249 +220,6 @@ class TestSnps(BaseSNPsTestCase):
             pos=[304103, 93431058, 624523],
             genotype=["AA", "AA", "AA"],
         )
-
-    def run_parsing_tests(self, file, source):
-        self.make_parsing_assertions(self.parse_file(file), source)
-        self.make_parsing_assertions(self.parse_bytes(file), source)
-
-    def parse_file(self, file):
-        return SNPs(file)
-
-    def parse_bytes(self, file):
-        with open(file, "rb") as f:
-            return SNPs(f.read())
-
-    def make_parsing_assertions(self, snps, source):
-        self.assertEqual(snps.source, source)
-        pd.testing.assert_frame_equal(snps.snps, self.generic_snps())
-
-    def test_snps_DNALand(self):
-        # https://dna.land/
-        self.run_parsing_tests("tests/input/DNALand.txt", "DNA.Land")
-
-    def test_snps_23andme(self):
-        # https://www.23andme.com
-        self.run_parsing_tests("tests/input/23andme.txt", "23andMe")
-
-    def test_snps_23andme_zip(self):
-        with atomic_write(
-            "tests/input/23andme.txt.zip", mode="wb", overwrite=True
-        ) as f:
-            with zipfile.ZipFile(f, "w") as f_zip:
-                # https://stackoverflow.com/a/16104667
-                f_zip.write("tests/input/23andme.txt", arcname="23andme.txt")
-        self.run_parsing_tests("tests/input/23andme.txt.zip", "23andMe")
-
-    def test_snps_ftdna(self):
-        # https://www.familytreedna.com
-        self.run_parsing_tests("tests/input/ftdna.csv", "FTDNA")
-
-    def test_snps_ftdna_gzip(self):
-        with open("tests/input/ftdna.csv", "rb") as f_in:
-            with atomic_write(
-                "tests/input/ftdna.csv.gz", mode="wb", overwrite=True
-            ) as f_out:
-                with gzip.open(f_out, "wb") as f_gzip:
-                    shutil.copyfileobj(f_in, f_gzip)
-        self.run_parsing_tests("tests/input/ftdna.csv.gz", "FTDNA")
-
-    def test_snps_ftdna_famfinder(self):
-        # https://www.familytreedna.com
-        self.run_parsing_tests("tests/input/ftdna_famfinder.csv", "FTDNA")
-
-    def test_snps_ancestry(self):
-        # https://www.ancestry.com
-        self.run_parsing_tests("tests/input/ancestry.txt", "AncestryDNA")
-
-    def test_snps_genes_for_good(self):
-        # https://genesforgood.sph.umich.edu/
-        self.run_parsing_tests("tests/input/genesforgood.txt", "GenesForGood")
-
-    def test_snps_myheritage(self):
-        # https://www.myheritage.com
-        self.run_parsing_tests("tests/input/myheritage.csv", "MyHeritage")
-
-    @staticmethod
-    def _setup_gsa_test():
-        # reset resource if already loaded
-        temp = SNPs()
-        temp._resources._gsa_resources = {}
-
-        create_dir("resources")
-
-        with open("tests/resources/gsa_rsid_map.txt", "rb") as f_in:
-            with atomic_write(
-                "resources/gsa_rsid_map.txt.gz", mode="wb", overwrite=True
-            ) as f_out:
-                with gzip.open(f_out, "wb") as f_gzip:
-                    shutil.copyfileobj(f_in, f_gzip)
-
-        with open("tests/resources/gsa_chrpos_map.txt", "rb") as f_in:
-            with atomic_write(
-                "resources/gsa_chrpos_map.txt.gz", mode="wb", overwrite=True
-            ) as f_out:
-                with gzip.open(f_out, "wb") as f_gzip:
-                    shutil.copyfileobj(f_in, f_gzip)
-
-    @staticmethod
-    def _teardown_gsa_test():
-        os.remove("resources/gsa_rsid_map.txt.gz")
-        os.remove("resources/gsa_chrpos_map.txt.gz")
-
-    def test_snps_codigo46(self):
-        # https://codigo46.com.mx
-        self._setup_gsa_test()
-        self.run_parsing_tests("tests/input/codigo46.txt", "Codigo46")
-        self._teardown_gsa_test()
-
-    def test_snps_sano(self):
-        # https://sanogenetics.com
-        self._setup_gsa_test()
-        self.run_parsing_tests("tests/input/sano.txt", "Sano")
-        self._teardown_gsa_test()
-
-    def test_snps_livingdna(self):
-        # https://livingdna.com
-        self.run_parsing_tests("tests/input/livingdna.csv", "LivingDNA")
-
-    def test_snps_mapmygenome(self):
-        # https://mapmygenome.in
-        self.run_parsing_tests("tests/input/mapmygenome.txt", "Mapmygenome")
-
-    def test_snps_vcf(self):
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        s = SNPs("tests/input/testvcf.vcf")
-        assert s.source == "vcf"
-        assert not s.unannotated_vcf
-        assert not s.phased
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
-
-    def test_snps_vcf_phased(self):
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        s = SNPs("tests/input/testvcf_phased.vcf")
-        assert s.source == "vcf"
-        assert not s.unannotated_vcf
-        assert s.phased
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
-
-    def test_snps_vcf_rsids(self):
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        rsids = ["rs1", "rs2"]
-        s = SNPs("tests/input/testvcf.vcf", rsids=rsids)
-        assert s.source == "vcf"
-        assert not s.unannotated_vcf
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf().loc[rsids])
-
-    def test_snps_vcf_gz(self):
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        with open("tests/input/testvcf.vcf", "rb") as f_in:
-            with atomic_write(
-                "tests/input/testvcf.vcf.gz", mode="wb", overwrite=True
-            ) as f_out:
-                with gzip.open(f_out, "wb") as f_gzip:
-                    shutil.copyfileobj(f_in, f_gzip)
-
-        s = SNPs("tests/input/testvcf.vcf.gz")
-        assert s.source == "vcf"
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
-
-    def test_snps_vcf_gz_rsids(self):
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        with open("tests/input/testvcf.vcf", "rb") as f_in:
-            with atomic_write(
-                "tests/input/testvcf.vcf.gz", mode="wb", overwrite=True
-            ) as f_out:
-                with gzip.open(f_out, "wb") as f_gzip:
-                    shutil.copyfileobj(f_in, f_gzip)
-
-        rsids = ["rs1", "rs2"]
-        s = SNPs("tests/input/testvcf.vcf.gz", rsids=rsids)
-        assert s.source == "vcf"
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf().loc[rsids])
-
-    def test_snps_unannotated_vcf(self):
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        s = SNPs("tests/input/unannotated_testvcf.vcf")
-        assert s.source == "vcf"
-        assert s.unannotated_vcf
-
-    def test_snps_not_phased(self):
-        s = SNPs("tests/input/generic.csv")
-        assert s.source == "generic"
-        assert not s.phased
-
-    def test_snps_vcf_buffer(self):
-        with open("tests/input/testvcf.vcf", "r") as f:
-            snps_vcf_buffer = SNPs(f.read().encode("utf-8"))
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        assert snps_vcf_buffer.source == "vcf"
-        pd.testing.assert_frame_equal(snps_vcf_buffer.snps, self.generic_snps_vcf())
-
-    def test_snps_vcf_buffer_rsids(self):
-        with open("tests/input/testvcf.vcf", "r") as f:
-            rsids = ["rs1", "rs2"]
-            df = SNPs(f.read().encode("utf-8"), rsids=rsids)
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        assert df.source == "vcf"
-        pd.testing.assert_frame_equal(df.snps, self.generic_snps_vcf().loc[rsids])
-
-    def test_snps_vcf_buffer_gz(self):
-        with open("tests/input/testvcf.vcf", "rb") as f_in:
-            with atomic_write(
-                "tests/input/testvcf.vcf.gz", mode="wb", overwrite=True
-            ) as f_out:
-                with gzip.open(f_out, "wb") as f_gzip:
-                    shutil.copyfileobj(f_in, f_gzip)
-
-        with open("tests/input/testvcf.vcf.gz", "rb") as f:
-            data = f.read()
-            s = SNPs(data)
-        os.remove("tests/input/testvcf.vcf.gz")
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        assert s.source == "vcf"
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
-
-    def test_snps_vcf_buffer_gz_rsids(self):
-        with open("tests/input/testvcf.vcf", "rb") as f_in:
-            with atomic_write(
-                "tests/input/testvcf.vcf.gz", mode="wb", overwrite=True
-            ) as f_out:
-                with gzip.open(f_out, "wb") as f_gzip:
-                    shutil.copyfileobj(f_in, f_gzip)
-
-        with open("tests/input/testvcf.vcf.gz", "rb") as f:
-            rsids = ["rs1", "rs2"]
-            data = f.read()
-            s = SNPs(data, rsids=rsids)
-        os.remove("tests/input/testvcf.vcf.gz")
-
-        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
-        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
-        # phased snps, and snps with missing rsID
-        assert s.source == "vcf"
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf().loc[rsids])
-
-    def test_source_generic(self):
-        s = SNPs("tests/input/NCBI36.csv")
-        assert s.source == "generic"
 
     def test_snps_no_snps(self):
         s = SNPs()
@@ -561,89 +299,6 @@ class TestSnps(BaseSNPsTestCase):
             chrom="1", pos_start=1, pos_max=249250621, pos_step=10000
         )
         assert s.sex == ""
-
-    def test_save_snps(self):
-        snps = SNPs("tests/input/GRCh37.csv")
-        assert os.path.relpath(snps.save_snps()) == "output/generic_GRCh37.csv"
-        s_saved = SNPs("output/generic_GRCh37.csv")
-        pd.testing.assert_frame_equal(s_saved.snps, self.snps_GRCh37())
-
-    def test_save_snps_bytes(self):
-        snps = SNPs("tests/input/GRCh37.csv")
-        assert os.path.relpath(snps.save_snps()) == "output/generic_GRCh37.csv"
-        with open("output/generic_GRCh37.csv", "rb") as f:
-            s_saved = SNPs(f.read())
-        pd.testing.assert_frame_equal(s_saved.snps, self.snps_GRCh37())
-
-    def test_save_snps_tsv(self):
-        snps = SNPs("tests/input/generic.csv")
-        assert (
-            os.path.relpath(snps.save_snps("generic.tsv", sep="\t"))
-            == "output/generic.tsv"
-        )
-        self.run_parsing_tests("output/generic.tsv", "generic")
-
-    def test_save_snps_vcf(self):
-        s = SNPs("tests/input/testvcf.vcf")
-
-        r = Resources()
-        r._reference_sequences["GRCh37"] = {}
-        with open("tests/input/generic.fa", "rb") as f_in:
-            with atomic_write(
-                "tests/input/generic.fa.gz", mode="wb", overwrite=True
-            ) as f_out:
-                with gzip.open(f_out, "wb") as f_gzip:
-                    shutil.copyfileobj(f_in, f_gzip)
-
-        seq = ReferenceSequence(ID="1", path="tests/input/generic.fa.gz")
-
-        r._reference_sequences["GRCh37"]["1"] = seq
-
-        assert os.path.relpath(s.save_snps(vcf=True)) == "output/vcf_GRCh37.vcf"
-        s = SNPs("output/vcf_GRCh37.vcf")
-        assert not s.phased
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
-
-    def test_save_snps_vcf_phased(self):
-        # read phased data
-        s = SNPs("tests/input/testvcf_phased.vcf")
-
-        # setup resource to use test FASTA reference sequence
-        r = Resources()
-        r._reference_sequences["GRCh37"] = {}
-        with open("tests/input/generic.fa", "rb") as f_in:
-            with atomic_write(
-                "tests/input/generic.fa.gz", mode="wb", overwrite=True
-            ) as f_out:
-                with gzip.open(f_out, "wb") as f_gzip:
-                    shutil.copyfileobj(f_in, f_gzip)
-
-        seq = ReferenceSequence(ID="1", path="tests/input/generic.fa.gz")
-
-        r._reference_sequences["GRCh37"]["1"] = seq
-
-        # save phased data to VCF
-        assert os.path.relpath(s.save_snps(vcf=True)) == "output/vcf_GRCh37.vcf"
-        # read saved VCF
-        s = SNPs("output/vcf_GRCh37.vcf")
-        assert s.phased
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
-
-    def test_save_snps_csv_phased(self):
-        # read phased data
-        s = SNPs("tests/input/testvcf_phased.vcf")
-        # save phased data to CSV
-        assert os.path.relpath(s.save_snps()) == "output/vcf_GRCh37.csv"
-        # read saved CSV
-        s = SNPs("output/vcf_GRCh37.csv")
-        assert s.phased
-        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
-
-    def test_save_snps_specify_file(self):
-        s = SNPs("tests/input/GRCh37.csv")
-        assert os.path.relpath(s.save_snps("snps.csv")) == "output/snps.csv"
-        s_saved = SNPs("output/snps.csv")
-        pd.testing.assert_frame_equal(s_saved.snps, self.snps_GRCh37())
 
     def test_save_snps_no_snps(self):
         s = SNPs()
